Revert "changed the requirements.txt so it only has the needed libraries"
This reverts commit 570d17e5b59e7b79462bc4686f6fd40daeb37b9a.

Going back to the old requirements.txt as the code seems to not run due to index problems even though no index changes show up in any of the commits. Weird thing is that
I have commited the code from the cluster, i.e. from the only place I was using it. Problem might come from the library updates even though I highly doubt it. Other solution might be
from me deleting one of the input images by accident. We'll see if the change in the requiremtns.txt solves the issue.